Move settings module to top level package, and leave 'models' subpackage for only serialized data models
