+++ requests_cache/__init__.py
@@ -15,6 +15,7 @@ try:
     from .patcher import *
     from .serializers import *
     from .session import *
+    from .settings import *
 # Log and ignore ImportErrors, if imported outside a virtualenv (e.g., just to check __version__)
 except ImportError as e:
     logger.warning(e)
