+++ requests_cache/models/__init__.py
@@ -7,7 +7,6 @@ from requests import PreparedRequest, Request, Response
 from .raw_response import CachedHTTPResponse
 from .request import CachedRequest
 from .response import CachedResponse, set_response_defaults
-from .settings import ALL_METHODS, CacheSettings, FilterCallback, KeyCallback, RequestSettings
 
 AnyResponse = Union[Response, CachedResponse]
 AnyRequest = Union[Request, PreparedRequest, CachedRequest]
